# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnyGPT API",
    "version": "1.0.0",
    "description": "AnyGPT API gateway for multiple model providers. This spec covers core endpoints and admin operations."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "summary": "Chat completions (OpenAI compatible)",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion"
          }
        }
      }
    },
...
    "/v4/v4/chat/completions": {
      "post": {
        "summary": "Groq chat completions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion"
          }
        }
      }
    },
...
    "/v6/v6/chat/completions": {
      "post": {
        "summary": "OpenRouter chat completions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  }
}